Refactoring back into src/perovskite_solar_cell_database
