+++ pyproject.toml
@@ -136,8 +136,8 @@ where = ["src"]
 perovskite_solar_cell = "perovskite_solar_cell_database:perovskite_solar_cell"
 perovskite_solar_cell_database_app = "perovskite_solar_cell_database.apps:perovskite_solar_cell_database_app"
 perovskite_composition = "perovskite_solar_cell_database:perovskite_composition"
-ion_parser = "perovskite_solar_cell_database:ion_parser"
+ion_parser = "perovskite_solar_cell_database.parsers:ion_parser"
 perovskite_ions_app = "perovskite_solar_cell_database.apps:perovskite_ions"
-perovskite_tandem_cell = "perovskite_tandem_database.schema_packages:tandem_solar_cell"
-perovskite_tandem_parser = "perovskite_tandem_database.parsers:tandem_parser"
+perovskite_tandem_cell = "perovskite_solar_cell_database.schema_packages:tandem_solar_cell"
+perovskite_tandem_parser = "perovskite_solar_cell_database.parsers:tandem_parser"
 solar_cell_app = "perovskite_solar_cell_database.apps:solar_cells"
