+++ src/perovskite_solar_cell_database/__init__.py
@@ -35,34 +35,3 @@ perovskite_composition = PerovskiteCompositionEntryPoint(
     name='PerovskiteComposition',
     description='Schema package defined for the perovskite compositions.',
 )
-
-
-class IonParserEntryPoint(ParserEntryPoint):
-    def load(self):
-        from perovskite_solar_cell_database.parser import IonParser
-
-        return IonParser(**self.dict())
-
-
-ion_parser = IonParserEntryPoint(
-    name='PerovskiteIonParser',
-    description='Parse excel files containing perovskite ions.',
-    mainfile_name_re=r'.+\.xlsx',
-    mainfile_mime_re='application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
-    mainfile_contents_dict={
-        'Sheet1': {
-            '__has_all_keys': [
-                'perovskite_site',
-                'abbreviation',
-                'molecular_formula',
-                'smiles',
-                'common_name',
-                'iupac_name',
-                'cas_number',
-                'source_compound_iupac_name',
-                'source_compound_smiles',
-                'source_compound_cas_number',
-            ]
-        },
-    },
-)
